Feat/student list

+++ test_website_publishment/lib/models/student.dart
@@ -1,11 +1,10 @@
 class Student {
   // Name of the student
-  String name; 
+  String name;
 
   // Age of the student
-  double? age; 
+  int? age;
 
   // Constructor
-  Student(this.name, double this.age);
-
+  Student(this.name, int this.age);
 }
